today I learnt that good todo comments can provide context
```https://sophiebits.com/2025/07/21/todos-arent-for-doing

But to me, a good TODO looks more like this:

// TODO: If the user triple-clicks this button, the click handler errors because [xyz]

Would this be the highest-priority thing on the backlog? Probably not.
Most users won’t end up triple-clicking that button.
But the // TODO: doesn’t need to be a plan to actually do something.
Instead, it’s a note about “here’s an edge case that wasn’t handled”
or a suggestion for a better structure that the author didn’t make
time to implement — it captures a little slice of the author’s brain
and gives a little window into the rich context they had at the time
they wrote the code.

```